Dense optical flow implementation
Implement an optical flow to show the movement in the images.
First implementation was done using the image frames, but after further consideration, the results vastly improve when using videos instead as inputs.
TO DO: update directory manager to pass the video directories and update dense_sequence function to operate with video format